+++ testing/provisioner_test.go
@@ -48,12 +48,12 @@ func (s *S) TestFakeAppRestart(c *gocheck.C) {
 func (s *S) TestFakeAppGetMemory(c *gocheck.C) {
 	app := NewFakeApp("sou", "otm", 0)
 	app.Memory = 100
-	c.Assert(app.GetMemory(), gocheck.Equals, 100)
+	c.Assert(app.GetMemory(), gocheck.Equals, int64(100))
 }
 
 func (s *S) TestFakeAppGetSwap(c *gocheck.C) {
 	app := NewFakeApp("sou", "otm", 0)
-	c.Assert(app.GetSwap(), gocheck.Equals, 0)
+	c.Assert(app.GetSwap(), gocheck.Equals, int64(0))
 }
 
 func (s *S) TestFakeAppSerializeEnvVars(c *gocheck.C) {
